날짜 별 조회

+++ src/main/java/com/travelplan/domain/plandetail/repository/PlanDetailRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface PlanDetailRepository extends JpaRepository<PlanDetail, Long> {
+public interface PlanDetailRepository extends JpaRepository<PlanDetail, Integer> {
 
     List<PlanDetail> findByPlan(Plan plan);
